⚡ Optimising home page game cards
Seperating the covers into a background and a foreground element so load SVG elements of the card cover faster than the background.

This also allows the SVG elements to be reused from the cache when you navigate to the project pages.

Also using webp compression for reduced file sizes and using a consistent 400x600 pixels resolution.

To reduce the amount of web requests that are made, I could combine the SVGs together using SVG sprites.